+++ Expenses/Model/Goal+CoreDataProperties.swift
@@ -2,7 +2,7 @@
 //  Goal+CoreDataProperties.swift
 //  Expenses
 //
-//  Created by Lana on 14/03/17.
+//  Created by Lana on 15/03/17.
 //  Copyright © 2017 L. All rights reserved.
 //
 
@@ -17,7 +17,7 @@ extension Goal {
     }
 
     @NSManaged public var date: NSDate?
-    @NSManaged public var total: Int16
+    @NSManaged public var total: Int64
     @NSManaged public var type: Int16
     @NSManaged public var category: Category?
     @NSManaged public var tag: Tag?
